P4ADEV-4599-broker-personalization-for-receipt-email: moved RemoteEmailTemplateRetrieverServiceImplTest into package connector and DownloadEmailTemplateClientImplTest into package connector/client; added interceptor x to RestTemplate in SignedUrlServiceImpl and DownloadEmailTemplateClientImpl; remove default for property mail.template.repo-base-url
