+++ src/main/java/it/gov/pagopa/payhub/activities/connector/emailtemplates/RemoteEmailTemplateRetrieverService.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.payhub.activities.service.email.remote.retriever;
+package it.gov.pagopa.payhub.activities.connector.emailtemplates;
 
 import it.gov.pagopa.payhub.activities.dto.email.EmailTemplate;
 import it.gov.pagopa.payhub.activities.enums.EmailTemplateName;
