+++ src/main/resources/config/application.yml
@@ -62,7 +62,7 @@ export-flow-files:
 mail:
   sender-address: "\${MAIL_SENDER_ADDRESS:piattaforma-unitaria@noreply.com}"
   template:
-    repo-base-url: "\${MAIL_TEMPLATE_REPO_BASE_URL:https://raw.githubusercontent.com/pagopa/p4pa-email-templates/refs/heads/main}"
+    repo-base-url: "\${MAIL_TEMPLATE_REPO_BASE_URL:}"
 
 folders:
   shared: "\${SHARED_FOLDER_ROOT:/shared}"
